requirements.txt
cx_Oracle==5.3
Django==1.11
django-environ==0.4.1
mysqlclient==1.3.10
